Added new exceptions, implement with services, added global handler exceptions

+++ backend/domain/services/lizard_service.py
@@ -1,5 +1,5 @@
 from backend.domain.entities.lizard import Lizard
-
+from backend.domain.exceptions.custom_exceptions import NotFoundError
 
 class LizardService:
 
@@ -7,16 +7,28 @@ class LizardService:
         self.lizard_repository = lizard_repository
 
     def get_lizards(self) -> list[Lizard]:
-        return self.lizard_repository.get_all_lizards()
+        found_lizard = self.lizard_repository.get_all_lizards()
+        if not found_lizard:
+            raise NotFoundError("No lizards found")
+        return found_lizard
 
     def get_lizard_by_id(self, lizard_id: str) -> Lizard | None:
-        return self.lizard_repository.get_lizard_by_id(lizard_id)
+        found_lizard = self.lizard_repository.get_lizard_by_id(lizard_id)
+        if not found_lizard:
+            raise NotFoundError("Lizard not found")
+        return found_lizard
 
     def create_lizard(self, lizard_data) -> Lizard:
         return self.lizard_repository.create_lizard(lizard_data)
 
     def update_lizard(self, lizard_id: str, lizard_data) -> Lizard | None:
-        return self.lizard_repository.update_lizard(lizard_id, lizard_data)
+        found_lizard = self.lizard_repository.update_lizard(lizard_id, lizard_data)
+        if not found_lizard:
+            raise NotFoundError("Lizard not found")
+        return found_lizard
 
     def delete_lizard(self, lizard_id: str) -> bool:
+        found_lizard = self.lizard_repository.get_lizard_by_id(lizard_id)
+        if not found_lizard:
+            raise NotFoundError("Lizard not found")
         return self.lizard_repository.delete_lizard(lizard_id)
